+++ src/main/kotlin/com/jjarappappa/imom/domain/information/domain/repository/InformationRepository.kt
@@ -1,7 +1,7 @@
 package com.jjarappappa.imom.domain.information.domain.repository
 
-import com.jjarappappa.imom.domain.information.Category
-import com.jjarappappa.imom.domain.information.Information
+import com.jjarappappa.imom.domain.information.domain.Category
+import com.jjarappappa.imom.domain.information.domain.Information
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface InformationRepository: JpaRepository<Information, Long> {
