+++ src/main/kotlin/com/jjarappappa/imom/domain/information/presentation/dto/FindInformationRequest.kt
@@ -1,6 +1,6 @@
 package com.jjarappappa.imom.domain.information.presentation.dto
 
-import com.jjarappappa.imom.domain.information.Category
+import com.jjarappappa.imom.domain.information.domain.Category
 
 data class FindInformationRequest(
     val category: Category,
